Fix overflow in Col2Im

+++ xchainer/native/col2im.cc
@@ -26,8 +26,8 @@ Array Col2Im(
         const StackVector<int64_t, kMaxNdim>& stride,
         const StackVector<int64_t, kMaxNdim>& pad,
         const StackVector<int64_t, kMaxNdim>& out_size) {
-    int8_t batch_size = col.shape()[0];
-    int8_t channels = col.shape()[1];
+    int64_t batch_size = col.shape()[0];
+    int64_t channels = col.shape()[1];
     auto ndim = static_cast<int8_t>(stride.size());
 
     Shape padded_shape{batch_size, channels};
